rev-378: refactor fx module to not depend on icon
