+++ modules/kb.core.fx/src/main/java/module-info.java
@@ -2,7 +2,6 @@ module kb.core.fx {
     requires kotlin.stdlib;
     requires javafx.controls;
     requires org.kordamp.iconli.core;
-    requires kb.core.icon;
 
     exports kb.core.fx;
 }
