--- modules/kb.core.fx/src/main/kotlin/kb/core/fx/IconUtils.kt
@@ -1,19 +1,0 @@
-package kb.core.fx
-
-import javafx.geometry.Pos
-import javafx.scene.layout.HBox
-import kb.core.icon.FontIcon
-import org.kordamp.ikonli.Ikon
-
-@FXKtDSL
-fun fontIcon(ic: Ikon, size: Int): FontIcon {
-    return FontIcon.of(ic, size)
-}
-
-@Suppress("unused")
-@FXKtDSL
-fun FontIcon.centerIn(width: Int): HBox = hbox {
-    add(this@centerIn)
-    prefWidth = width.toDouble()
-    alignment = Pos.CENTER
-}
