+++ modules/kb.core.fx/src/main/kotlin/kb/core/fx/MenuUtils.kt
@@ -28,18 +28,6 @@ fun Menu.name(name: String) {
     text = name
 }
 
-@FXKtDSL
-fun Menu.icon(icon: Ikon, iconSize: Int) {
-    graphic = fontIcon(icon, iconSize)
-}
-
-@FXKtDSL
-fun MenuItem.icon(icon: Ikon, iconSize: Int) {
-    graphic = fontIcon(icon, iconSize)
-    // this is so that icons are moved away from the text
-    graphic.translateX = -2.0
-}
-
 typealias Combination = KeyCodeCombination
 
 @FXKtDSL
