+++ modules/kb.core.icon/icon.gradle.kts
@@ -1,11 +1,12 @@
 plugins {
     java
     id("org.openjfx.javafxplugin")
+    kotlin("jvm")
 }
 
-javafx {
-    modules = listOf("javafx.controls")
-}
+//javafx {
+//    modules = listOf("javafx.controls")
+//}
 
 dependencies {
     implementation(group = "org.kordamp.ikonli", name = "ikonli-core", version = "11.3.4")
