+++ modules/kb.core.view/src/main/kotlin/kb/core/view/TBAView.kt
@@ -9,6 +9,7 @@ import javafx.stage.Stage
 import javafx.stage.StageStyle
 import javafx.stage.Window
 import kb.core.fx.*
+import kb.core.icon.fontIcon
 import org.kordamp.ikonli.materialdesign.MaterialDesign
 
 @Suppress("MemberVisibilityCanBePrivate")
